Library admin refactor (#89097)
## About The Pull Request
Removes uselocalstate from library admin. Did some stylistic changes as
per usual, broke the code into multiple files
## Why It's Good For The Game
useLocalstate is deprecated, I want it out of the codebase
## Changelog
N/A
